+++ documentation-site/src/pages/demos/easing-functions/_create-easing-rows.ts
@@ -115,6 +115,7 @@ export async function createEasingRows(
   );
 
   const spriteTemplate = createImageSprite(whiteImage, renderContext, {
+    pixelsPerUnit: 1,
     layer: renderLayer,
   });